3.0.3

+++ src/SFtp.cc
@@ -1796,9 +1796,9 @@ FileInfo *SFtp::MakeFileInfo(const NameAttrs *na)
    if(a->flags&SSH_FILEXFER_ATTR_UIDGID)
    {
       char id[12];
-      sprintf(id,"%u",a->uid);
+      sprintf(id,"%u",(unsigned)a->uid);
       fi->SetUser(id);
-      sprintf(id,"%u",a->gid);
+      sprintf(id,"%u",(unsigned)a->gid);
       fi->SetGroup(id);
    }
    if(a->flags&SSH_FILEXFER_ATTR_OWNERGROUP)
